Do not filter headers
