+++ components/gateway/cmd/main.go
@@ -1,13 +1,10 @@
 package main
 
 import (
+	"github.com/kyma-incubator/compass/components/gateway/pkg/proxy"
+	"github.com/pkg/errors"
 	"log"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"strings"
-
-	"github.com/pkg/errors"
 
 	"github.com/gorilla/mux"
 	"github.com/vrischmann/envconfig"
@@ -25,25 +22,13 @@ func main() {
 	err := envconfig.InitWithPrefix(&cfg, "APP")
 	exitOnError(err, "Error while loading app config")
 
-	directorUrl, err := url.Parse(cfg.DirectorOrigin)
-	exitOnError(err, "Error while parsing Director URL")
+	directorProxy, err := proxy.New(cfg.DirectorOrigin)
+	exitOnError(err, "Error while initializing proxy")
 
 	log.Printf("Proxying requests to Director: %s\n", cfg.DirectorOrigin)
 
-	directorProxy := httputil.ReverseProxy{
-		Director: func(r *http.Request) {
-			r.URL.Scheme = directorUrl.Scheme
-			r.URL.Host = directorUrl.Host
-			r.URL.Path = singleJoiningSlash(directorUrl.Path, r.URL.Path)
-			r.Header = http.Header{
-				"Content-Type": []string{"application/json"},
-			}
-			r.Host = directorUrl.Host
-		},
-	}
-
 	router := mux.NewRouter()
-	router.HandleFunc("/", directorProxy.ServeHTTP)        // GraphQL Playground
+	router.HandleFunc("/", directorProxy.ServeHTTP) // GraphQL Playground
 	router.HandleFunc("/graphql", directorProxy.ServeHTTP) // GraphQL API Endpoint
 
 	router.HandleFunc("/healthz", func(writer http.ResponseWriter, request *http.Request) {
@@ -67,16 +52,4 @@ func exitOnError(err error, context string) {
 		wrappedError := errors.Wrap(err, context)
 		log.Fatal(wrappedError)
 	}
-}
-
-func singleJoiningSlash(a, b string) string {
-	aslash := strings.HasSuffix(a, "/")
-	bslash := strings.HasPrefix(b, "/")
-	switch {
-	case aslash && bslash:
-		return a + b[1:]
-	case !aslash && !bslash:
-		return a + "/" + b
-	}
-	return a + b
 }
